запросы для account

+++ src/main/java/com/example/bankcards/repository/TransactionJpaRepository.java
@@ -3,6 +3,6 @@ package com.example.bankcards.repository;
 import com.example.bankcards.entity.TransactionJpa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionJpaRepository extends JpaRepository<TransactionJpa, Integer> {
+public interface TransactionJpaRepository extends JpaRepository<TransactionJpa, Long> {
 
 }
